remove bower, remove localstorageservice, simplify project
The project had a dependency on angular-local-storage which caused some
problems in cases where the type of storage should depend on the type of
session. The storage type can now be changed at any time from
localStorage to sessionStorage. As we are generally using npm for all
dependency installs at vokal now, the bower related stuff has been
removed, including most of the build process.